requirements.txt
pygame==1.9.4
Pillow==5.3.0
